requirements.txt
requests
cymon
beautifulsoup4
maxminddb-geolite2
flask